+++ pkg/gen/input/precommit/internal/file/error.go
@@ -1,0 +1,7 @@
+package file
+
+import "github.com/giantswarm/microerror"
+
+var executionFailedError = &microerror.Error{
+	Kind: "executionFailedError",
+}