+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `devctl gen precommit`: Add required `--repo-name` flag and automatically detect helm charts in `helm/` directory to generate chart-specific configuration sections.
+
 ## [7.33.0] - 2026-02-06
 
 ### Removed